Add archive password to backup settings
A new archive password attribute has been added in the BackupSetting model, controller, and request. This password is used to secure the backup archives. Checks have been added to ensure any updates to the password are registered and saved, and comprehensive tests for these changes have been added as well.