Add ordering functionality to base repository
New "Order" related interfaces and abstractions have been introduced, and the base repository has been updated to include ordering functionality. With this update, the base repository is not only responsible for applying the criteria but also for applying orders, depending on whether or not the 'skipOrders' property is set. A new trait 'ModelFilterableFields' is included to separate models based on filterable fields.