Add initial recovery support
The update introduces a new feature: the ability to perform an initial recovery of the system. The provider has been extended with a new gate definition for the initial recovery. Additionally, a `RecoveryController` and corresponding `InitialRecoveryRequest` have been created to manage the recovery and validation process respectively. Also, a new route has been added linking to the recovery function.